requirements.txt
google-generativeai
PyPDF2
python-dotenv
sentence-transformers
faiss-cpu
httpx
python-docx
pdfminer.six
waitress
fastapi
uvicorn
python-jose
requests
python-multipart